Rename 'generic' to 'workflow' to reflect external naming
