+++ tests/integration/workflow/test_test_suite.py
@@ -25,13 +25,13 @@ from kolena.detection import TestSuite as DetectionTestSuite
 from kolena.errors import NameConflictError
 from kolena.errors import WorkflowMismatchError
 from kolena.workflow import define_workflow
-from tests.integration.generic.dummy import DUMMY_WORKFLOW
-from tests.integration.generic.dummy import DummyGroundTruth
-from tests.integration.generic.dummy import DummyInference
-from tests.integration.generic.dummy import DummyTestSample
-from tests.integration.generic.dummy import TestCase
-from tests.integration.generic.dummy import TestSuite
 from tests.integration.helper import with_test_prefix
+from tests.integration.workflow.dummy import DUMMY_WORKFLOW
+from tests.integration.workflow.dummy import DummyGroundTruth
+from tests.integration.workflow.dummy import DummyInference
+from tests.integration.workflow.dummy import DummyTestSample
+from tests.integration.workflow.dummy import TestCase
+from tests.integration.workflow.dummy import TestSuite
 
 
 @pytest.fixture(scope="module")
